cleared warnings about require cycle + fixed discovery read article back leads to home bug
