+++ src/components/shared/CustomHeader.js
@@ -5,13 +5,11 @@ import {
   Image,
   StatusBar,
   SafeAreaView,
-  Text,
   Platform,
   Dimensions,
 } from 'react-native'
 import { TouchableOpacity } from 'react-native-gesture-handler'
-import { PublicationEnum } from '../../../NavigationController'
-import { PublicationPrimaryColorRgba } from '../../utils/branding'
+import { PublicationEnum } from '../../utils/constants'
 const DP_LOGO_WHITE = require('../../static/logos/dp-logo-small-white.png')
 const DP_LOGO_RED = require('../../static/logos/dp-logo-small-red.png')
 // half the height of the header
