+++ src/screens/ScreenWithDefaultParams.js
@@ -1,0 +1,8 @@
+import React, { Component } from 'react'
+export const ScreenWithDefaultParams = (Comp, defaultParams) => {
+  return class extends Component {
+    render() {
+      return <Comp {...this.props} screenProps={defaultParams} />
+    }
+  }
+}
