+++ src/screens/index.js
@@ -8,3 +8,4 @@ export * from './NotificationScreen'
 export * from './ManageFeedScreen'
 export * from './WebViewScreen'
 export * from './SectionScreen'
+export * from './ScreenWithDefaultParams'
